refactor: mvc 디자인 패턴 기반으로 파일 분리
기존 index.js에 모여있었던 코드를, 유지보수 성능 향상을 위해 mvc 패턴을 기반으로 파일을 나누었다.
