+++ docs/README.md
@@ -6,13 +6,19 @@
   - [예외처리] 5자를 초과한 것이 있을 때 
   - [예외처리] 중복되는 이름이 있을 때 
   - [예외처리] 공백을 입력했을 때 
-- [x] 유효하다면, 시도 횟수를 입력받는 form을 렌더링한다. 
+- [x] 유효하다면
+  - [x] 시도 횟수를 입력받는 form을 렌더링한다. 
+  - [x] 자동차 이름을 입력받는 input은 비활성화 한다. 
+  - [x] 입력한 자동차 리스트를 객체로 변환하고, 여러개일경우 배열로 만든다. 
 
 **2. 시도 횟수를 입력받는 기능**
 - [x] 사용자는 시도할 횟수를 입력하고, 엔터 혹은 확인버튼을 클릭한다. 
 - [x] 입력받은 값이 유효하지 않으면, alert를 이용해 메시지를 보여주고, 다시 입력할 수 있게 한다. 
   - [예외처리] 공백인 경우 
   - [예외처리] 0 이하인 경우
+- [x] 유효하다면  
+  - [x] 시도 횟수를 입력받는 input창을 비활성화 한다. 
+  - [x] 게임을 시작한다. 
 
 **3. 자동차를 전진 또는 정지할지 결정하는 기능** 
 - [x] 입력한 자동차들을 객체로 생성한다. (객체들은 배열에 넣는다. )
@@ -33,10 +39,20 @@
   - [x] 우승자가 여러명일 경우, 쉼표(,)를 이용하여 구분하여 출력한다. 
   - [x] id가 'racing-winners'인 span태그를 만들고, 이 태그 내에 최종 우승자를 출력한다. 
 
+## 리팩터링 
+- [x] 클래스 분리 
+- [x] 파일분리(디자인 패턴?)-mvc 패턴 
+- [x] input 렌더링하는것 리팩토링 
+- [ ] 사용성 고민: 자동 포커스 기능을 통해 사용자가 프로그램 시작되면 가장 먼저해야할일 암시
+- [ ] alert는 void형이기 때문에, 삼항연산자 지양 
+- [ ] trim() 해주는 것 따로 함수 만들기(함수는 기능 1개만)
+- [ ] ,,으로 입력됐을 때 예외 
+- [ ] hidden없애고 id로 하기 
+- [ ] 보여주고 보여주지 않는것을 하나의 파일로 정리해보기 (매개변수에 태그넣으면, 그게 보이고 안보이게)
 
 ## ✅ 요구 사항
-- [ ] 주어진 index.html에 html 엘리먼트를 직접 추가하거나 기존의 html 엘리먼트를 임의로 삭제하지 않는다. id와 같은 선택자를 추가하는 작업만 가능하다.
-- [ ] 다음과 같이 Car 객체를 만들고, new 를 이용해 인스턴스를 만들어 사용한다.
+- [x] 주어진 index.html에 html 엘리먼트를 직접 추가하거나 기존의 html 엘리먼트를 임의로 삭제하지 않는다. id와 같은 선택자를 추가하는 작업만 가능하다.
+- [x] 다음과 같이 Car 객체를 만들고, new 를 이용해 인스턴스를 만들어 사용한다.
   ```javascript
   function Car(name) {
   this.name = name;
@@ -49,14 +65,14 @@
   - 레이싱 횟수을 제출하는 button 태그는 `racing-count-submit` id값을 가진다.
   - 최종 우승자를 출력하는 span 태그는 `racing-winners` id값을 가진다.
     - 예) <span id="racing-winners">poco,park,jun</span>
-- [ ] 전진하는 조건을 판단하기 위한 랜덤 값은 MissionUtils 라이브러리의 Random.pickNumberInRange를 사용해 구한다. MissionUtils 라이브러리 스크립트는 index.html에 이미 포함되어 전역 객체에 추가되어 있으므로, 따로 import 하지 않아도 구현 코드 어디에서든 사용할 수 있다.
+- [x] 전진하는 조건을 판단하기 위한 랜덤 값은 MissionUtils 라이브러리의 Random.pickNumberInRange를 사용해 구한다. MissionUtils 라이브러리 스크립트는 index.html에 이미 포함되어 전역 객체에 추가되어 있으므로, 따로 import 하지 않아도 구현 코드 어디에서든 사용할 수 있다.
 
-- [ ] 외부 라이브러리(jQuery, Lodash 등)를 사용하지 않고, 순수 Vanilla JS로만 구현한다.
-- [ ] 자바스크립트 코드 컨벤션을 지키면서 프로그래밍 한다. 정답이 없으므로, 다양한 컨벤션을 비교해보며 스스로 더 적절해보이는 컨벤션을 자율적으로 선택한다.
-- [ ] indent(인덴트, 들여쓰기) depth를 3이 넘지 않도록 구현한다. 2까지만 허용한다. 예를 들어 while문 안에 if문이 있으면 들여쓰기는 2이다.
+- [x] 외부 라이브러리(jQuery, Lodash 등)를 사용하지 않고, 순수 Vanilla JS로만 구현한다.
+- [x] 자바스크립트 코드 컨벤션을 지키면서 프로그래밍 한다. 정답이 없으므로, 다양한 컨벤션을 비교해보며 스스로 더 적절해보이는 컨벤션을 자율적으로 선택한다.
+- [] indent(인덴트, 들여쓰기) depth를 3이 넘지 않도록 구현한다. 2까지만 허용한다. 예를 들어 while문 안에 if문이 있으면 들여쓰기는 2이다.
 - [ ] 함수(또는 메소드)가 한 가지 일만 하도록 최대한 작게 만들어라.
-- [ ] 변수 선언시 var 를 사용하지 않는다. const 와 let 을 사용한다.
-- [ ] import 문을 이용해 스크립트를 모듈화하고 불러올 수 있게 만든다.
+- [x] 변수 선언시 var 를 사용하지 않는다. const 와 let 을 사용한다.
+- [x] import 문을 이용해 스크립트를 모듈화하고 불러올 수 있게 만든다.
 - [ ] 함수(또는 메소드)의 길이가 15라인을 넘어가지 않도록 구현한다.
 - [ ] 함수(또는 메소드)가 한 가지 일만 잘 하도록 구현한다.
 - [ ] 테스트에서 all pass를 한다. (`npm run test`)
@@ -80,18 +96,3 @@
 - [ ] 최종 제출 코드에서 EOL(End Of Line) 설정을 확인하라. 환경에 따라 의도한 바와 다르게 개행 문자가 처리되지 않도록 EOL 설정을 확인한다. 
 - [ ] 최종 제출 코드에서 불필요한 console.log를 남기지 않는다. 
 - [ ] var 대신 const, let을 사용하라. 
-
-## 리팩터링 
-- [ ] 클래스 분리 
-- [ ] 파일분리(디자인 패턴?)
-- [ ] else if, else 덜쓰기 (js else if 대신 검색~, switch?)
-- [ ] 모듈화 (mvc 패턴, react 컴포넌트화)
-- [ ] 사용성 고민: 자동 포커스 기능을 통해 사용자가 프로그램 시작되면 가장 먼저해야할일 암시
-- [ ] input 렌더링하는것 리팩토링 
-- [ ] 예외처리 더 명시적으로 
-- [ ] alert는 void형이기 때문에, 삼항연산자 지양 
-- [ ] 읽기 쉬운 코드 만들기 (클린코드 ref 참고)
-- [ ] trim() 해주는 것 따로 함수 만들기(함수는 기능 1개만)
-- [ ] html script 링크 추가 (왜 안해도 실행 되는지...??)
-- [ ] ,,으로 입력됐을 때 예외 
-- [ ] submit vs click
