+++ index.html
@@ -25,6 +25,6 @@
       </form>
       <h4>📄 실행 결과</h4>
     </div>
-    <script type="module" src="src/index.js"></script>
+    <script type="module" src="src/js/index.js"></script>
   </body>
 </html>
