+++ src/js/index.js
@@ -1,0 +1,16 @@
+import $ from './utils/dom.js';
+import handleCarNameInput from './handlers/handleCarNameInput.js';
+import handleRacingCountInput from './handlers/handleRacingCountInput.js';
+
+function RacingCarGame() {
+  $('#car-names-submit').addEventListener('click', e => {
+    e.preventDefault();
+    handleCarNameInput();
+  });
+  $('#racing-count-submit').addEventListener('click', e => {
+    e.preventDefault();
+    handleRacingCountInput();
+  });
+}
+
+new RacingCarGame();
