+++ src/js/models/Car.js
@@ -1,0 +1,4 @@
+export default function Car(name) {
+  this.name = name;
+  this.move = 0;
+}
