+++ src/js/views/renderCountInput.js
@@ -1,0 +1,8 @@
+import $ from '../utils/dom.js';
+
+const renderCountInput = () => {
+  $('#racing-count-title').className = 'visible';
+  $('#racing-count-form').className = 'visible';
+};
+
+export default renderCountInput;
