+++ src/js/views/resetCarInput.js
@@ -1,0 +1,7 @@
+import $ from '../utils/dom.js';
+
+const resetCarInput = () => {
+  $('#car-names-input').value = '';
+};
+
+export default resetCarInput;
