+++ src/js/views/resetCountInput.js
@@ -1,0 +1,5 @@
+const resetCountInput = () => {
+  $('#racing-count-input').value = '';
+};
+
+export default resetCountInput;
